Collapse unit tests
Collapse the unit tests in `ExecutablePathValueProviderTests` and `LaunchTargetValueProviderTests` into the `ActiveLaunchProfilesValueProviderTests` class. The tests are very similar, and each file only had three tests, so it makes sense to merge them in order to make it easier to extract out some common test infrastructure in a later commit.
