+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/ProjectSystem/Properties/InterceptedProjectProperties/ActiveLaunchProfilesValueProviderTests.cs
@@ -1,6 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
 using System.Threading.Tasks;
+using Microsoft.VisualStudio.ProjectSystem.Debug;
 using Moq;
 using Xunit;
 
@@ -44,6 +45,151 @@ namespace Microsoft.VisualStudio.ProjectSystem.Properties
 
             Assert.Null(result);
             Assert.Equal(expected: "Delta", actual: activeProfileName);
+        }
+
+        [Fact]
+        public async Task ExecutablePath_OnGetEvaluatedPropertyValueAsync_GetsExecutableFromActiveProfile()
+        {
+            string activeProfileName = "Alpha";
+            string activeProfileLaunchTarget = "AlphaCommand";
+            string activeProfileExecutablePath = @"C:\user\bin\alpha.exe";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget,
+                ExecutablePath = activeProfileExecutablePath
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(activeProfileName, new[] { profile.ToLaunchProfile() });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new ExecutablePathValueProvider(project, settingsProvider, threadingService);
+
+            var actualValue = await launchProfileProvider.OnGetEvaluatedPropertyValueAsync(string.Empty, Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: activeProfileExecutablePath, actual: actualValue);
+        }
+
+        [Fact]
+        public async Task ExecutablePath_OnGetUnevaluatedPropertyValueAsync_GetsExecutableFromActiveProfile()
+        {
+            string activeProfileName = "Beta";
+            string activeProfileLaunchTarget = "BetaCommand";
+            string activeProfileExecutablePath = @"C:\user\bin\beta.exe";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget,
+                ExecutablePath = activeProfileExecutablePath
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(activeProfileName, new[] { profile.ToLaunchProfile() });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new ExecutablePathValueProvider(project, settingsProvider, threadingService);
+
+            var actualValue = await launchProfileProvider.OnGetUnevaluatedPropertyValueAsync(string.Empty, Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: activeProfileExecutablePath, actual: actualValue);
+        }
+
+        [Fact]
+        public async Task ExecutablePath_OnSetPropertyValueAsync_SetsTargetInActiveProfile()
+        {
+            string activeProfileName = "Gamma";
+            string activeProfileLaunchTarget = "GammaCommand";
+            string activeProfileExecutablePath = @"C:\user\bin\gamma.exe";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget,
+                ExecutablePath = activeProfileExecutablePath
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(
+                activeProfileName,
+                new[] { profile.ToLaunchProfile() },
+                updateLaunchSettingsCallback: s =>
+                {
+                    activeProfileExecutablePath = s.ActiveProfile!.ExecutablePath;
+                });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new ExecutablePathValueProvider(project, settingsProvider, threadingService);
+
+            await launchProfileProvider.OnSetPropertyValueAsync(@"C:\user\bin\delta.exe", Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: @"C:\user\bin\delta.exe", actual: activeProfileExecutablePath);
+        }
+
+        [Fact]
+        public async Task LaunchTarget_OnGetEvaluatedPropertyValueAsync_GetsTargetFromActiveProfile()
+        {
+            string activeProfileName = "Alpha";
+            string activeProfileLaunchTarget = "AlphaCommand";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(activeProfileName, new[] { profile.ToLaunchProfile() });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new LaunchTargetValueProvider(project, settingsProvider, threadingService);
+
+            var actualValue = await launchProfileProvider.OnGetEvaluatedPropertyValueAsync(string.Empty, Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: activeProfileLaunchTarget, actual: actualValue);
+
+        }
+
+        [Fact]
+        public async Task LaunchTarget_OnGetUnevaluatedPropertyValueAsync_GetsTargetFromActiveProfile()
+        {
+            string activeProfileName = "Beta";
+            string activeProfileLaunchTarget = "BetaCommand";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(activeProfileName, new[] { profile.ToLaunchProfile() });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new LaunchTargetValueProvider(project, settingsProvider, threadingService);
+
+            var actualValue = await launchProfileProvider.OnGetEvaluatedPropertyValueAsync(string.Empty, Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: activeProfileLaunchTarget, actual: actualValue);
+        }
+
+        [Fact]
+        public async Task LaunchTarget_OnSetPropertyValueAsync_SetsTargetInActiveProfile()
+        {
+            string activeProfileName = "Gamma";
+            string activeProfileLaunchTarget = "GammaCommand";
+            var profile = new WritableLaunchProfile
+            {
+                Name = activeProfileName,
+                CommandName = activeProfileLaunchTarget
+            };
+            var settingsProvider = ILaunchSettingsProviderFactory.Create(
+                activeProfileName,
+                new[] { profile.ToLaunchProfile() },
+                updateLaunchSettingsCallback: s =>
+                {
+                    activeProfileLaunchTarget = s.ActiveProfile!.CommandName;
+                });
+
+            var project = UnconfiguredProjectFactory.Create();
+            var threadingService = IProjectThreadingServiceFactory.Create();
+            var launchProfileProvider = new LaunchTargetValueProvider(project, settingsProvider, threadingService);
+
+            await launchProfileProvider.OnSetPropertyValueAsync("NewCommand", Mock.Of<IProjectProperties>());
+
+            Assert.Equal(expected: "NewCommand", actual: activeProfileLaunchTarget);
         }
     }
 }
